# anyharness/sdk/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AnyHarness Runtime API",
    "description": "Local sidecar runtime for managing coding agent installations, credentials, and ACP sessions.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "list_workspaces",
        "responses": {
          "200": {
            "description": "List workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "create_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveWorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/resolve": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "resolve_workspace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveWorkspaceFromPathRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resolved workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveWorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid path",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/worktrees": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "create_worktree",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorktreeWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Created worktree workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateWorktreeWorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Source workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "get_workspace",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{workspace_id}/detect-setup": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "detect_project_setup",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detected project setup hints",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectProjectSetupResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/display-name": {
      "patch": {
        "tags": [
          "workspaces"
        ],
        "operationId": "update_workspace_display_name",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceDisplayNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated workspace display name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid display name",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{workspace_id}/session-launch": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "get_workspace_session_launch_catalog",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace session launch catalog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceSessionLaunchCatalog"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/workspaces/{workspace_id}/setup-rerun": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "rerun_setup",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setup execution restarted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSetupStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "No setup script configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/setup-start": {
      "post": {
        "tags": [
          "workspaces"
        ],
        "operationId": "start_setup",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartWorkspaceSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Setup execution started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSetupStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid setup command",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/setup-status": {
      "get": {
        "tags": [
          "workspaces"
        ],
        "operationId": "get_setup_status",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "description": "Workspace ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setup execution status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSetupStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "No setup execution found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "CreateWorktreeWorkspaceRequest": {
        "type": "object",
        "required": [
          "repoRootId",
          "targetPath",
          "newBranchName"
        ],
        "properties": {
          "baseBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "newBranchName": {
            "type": "string"
          },
          "repoRootId": {
            "type": "string"
          },
          "setupScript": {
            "type": [
              "string",
              "null"
            ]
          },
          "targetPath": {
            "type": "string"
          }
        }
      },
      "CreateWorktreeWorkspaceResponse": {
        "type": "object",
        "required": [
          "workspace"
        ],
        "properties": {
          "setupScript": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SetupScriptExecution"
              }
            ]
          },
          "workspace": {
            "$ref": "#/components/schemas/Workspace"
          }
        }
      },
...
      "DetectProjectSetupResponse": {
        "type": "object",
        "required": [
          "hints"
        ],
        "properties": {
          "hints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SetupHint"
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ResolveWorkspaceFromPathRequest": {
        "type": "object",
        "required": [
          "path"
        ],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "ResolveWorkspaceResponse": {
        "type": "object",
        "required": [
          "repoRoot",
          "workspace"
        ],
        "properties": {
          "repoRoot": {
            "$ref": "#/components/schemas/RepoRoot"
          },
          "workspace": {
            "$ref": "#/components/schemas/Workspace"
          }
        }
      },
...
      "Workspace": {
        "type": "object",
        "required": [
          "id",
          "kind",
          "repoRootId",
          "path",
          "surface",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "currentBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "displayName": {
            "type": [
              "string",
              "null"
            ]
          },
          "executionSummary": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/WorkspaceExecutionSummary"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/WorkspaceKind"
          },
          "originalBranch": {
            "type": [
              "string",
              "null"
            ]
          },
          "path": {
            "type": "string"
          },
          "repoRootId": {
            "type": "string"
          },
          "surface": {
            "$ref": "#/components/schemas/WorkspaceSurface"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "WorkspaceKind": {
        "type": "string",
        "enum": [
          "worktree",
          "local"
        ]
      },
...
    }
  }
}